+++ src/decays/H_SM_decays/decay_H_to_WW.cpp
@@ -10,11 +10,9 @@ double CLASSNAME::get_partial_width<H, conj<W>::type, W>(
 
    const double mHOS = context.physical_mass<H>(indexIn);
    const double mWOS = context.physical_mass<W>(indexOut1);
-   const double mWDR = context.mass<W>(indexOut1);
    const double x = Sqr(mWOS/mHOS);
    double res;
 
-   // 4-body decay for mH < mW not implemented for a moment
    if (4.*x > 1.0) {
 
       // integrand