+++ src/decays/H_SM_decays/decay_H_to_dqdq.cpp
@@ -69,14 +69,15 @@ double CLASSNAME::get_partial_width<H,bar<dq>::type,dq>(
 
    switch (include_higher_order_corrections) {
       case SM_higher_order_corrections::enable: {
-         double deltaqq_QCD_OS = 0.;
          const int Nf = number_of_active_flavours(qedqcd, mHOS);
          double alpha_s_red;
+         double Y_conversion = 1.;
          switch (Nf) {
             case 5: {
                auto qedqcd_ = qedqcd;
                qedqcd_.to(mHOS);
                alpha_s_red = qedqcd_.displayAlpha(softsusy::ALPHAS)/Pi;
+               Y_conversion = Sqr(sm_down_quark_masses(qedqcd_, indexOut1.at(0))/mdqDR);
                break;
             }
             case 6:
@@ -96,6 +97,9 @@ double CLASSNAME::get_partial_width<H,bar<dq>::type,dq>(
             2.*(1. - 10.*xDR)/(1-4.*xDR)*(4./3. - std::log(xDR))*alpha_s_red +
             4./3.*alpha_s_red*calc_DeltaH(betaDR);
 
+         const double deltaqq_QCDxQED_DR =
+            (691/24. - 6*zeta3 - Sqr(Pi))*Sqr(dq::electric_charge)*alpha_red*alpha_s_red;
+
          const double deltaqq_QCD_OS_S =
             4./3. * alpha_s_red * calc_DeltaH(betaOS);
 
@@ -113,7 +117,7 @@ double CLASSNAME::get_partial_width<H,bar<dq>::type,dq>(
             // eq. 28 of hep-ph/9505358
             const auto Httbar_S = 0.5*(Httbar.left() + Httbar.right());
             const auto gtHoVEV = Httbar_S/context.mass<uq>({2});
-            deltaPhi2_S = Sqr(alpha_s_red) * std::real(gtHoVEV/gbHoVEV) * (1.57 - 2.0/3.0*lt + 1.0/9.0*Sqr(lq));
+            deltaPhi2_S = Sqr(alpha_s_red) * std::real(gtHoVEV/gbHoVEV) * (8/3. - Sqr(Pi/3.) - 2.0/3.0*lt + 1.0/9.0*Sqr(lq));
          }
 
          double deltaqq_QCD_OS_P = 0.;
@@ -136,12 +140,12 @@ double CLASSNAME::get_partial_width<H,bar<dq>::type,dq>(
             if (!is_zero(gbHoVEV_P)) {
                const auto Httbar_P = 0.5*(Httbar.right() - Httbar.left());
                const auto gtHoVEV_P = Httbar_P/context.mass<uq>({2});
-               deltaPhi2_P = Sqr(alpha_s_red) * std::real(gtHoVEV_P/gbHoVEV_P) * (3.83 - lt + 1.0/6.0*Sqr(lq));
+               deltaPhi2_P = Sqr(alpha_s_red) * std::real(gtHoVEV_P/gbHoVEV_P) * (23/6. - lt + 1.0/6.0*Sqr(lq));
             }
          }
 
-         amp2DR_S *= 1. + deltaqq_QCD_DR_S + deltaqq_QED_DR + deltaPhi2_S;
-         amp2DR_P *= 1. + deltaqq_QCD_DR_P + deltaqq_QED_DR + deltaPhi2_P;
+         amp2DR_S *= Y_conversion*(1. + deltaqq_QCD_DR_S + deltaqq_QED_DR + deltaqq_QCDxQED_DR + deltaPhi2_S);
+         amp2DR_P *= Y_conversion*(1. + deltaqq_QCD_DR_P + deltaqq_QED_DR + deltaqq_QCDxQED_DR + deltaPhi2_P);
          amp2OS_S *= 1. + deltaqq_QCD_OS_S + deltaqq_QED_OS_S;
          amp2OS_P *= 1. + deltaqq_QCD_OS_P + deltaqq_QED_OS_P;
          break;
@@ -149,7 +153,7 @@ double CLASSNAME::get_partial_width<H,bar<dq>::type,dq>(
       case SM_higher_order_corrections::disable:
          break;
       default:
-         WARNING("Unhandled option in H->ddbar decay");
+         throw std::runtime_error("Unhandled option in H->ddbar decay");
    }
 
    // low x limit