+++ CHANGES.rst
@@ -4,7 +4,7 @@ FlexibleSUSY 2.6.0 [not released yet]
 New features
 ------------
 
-* Added FlexibleDecays.
+* Added FlexibleDecay.
 
   Thanks to ME! And maybe Dylan.
 
@@ -20,6 +20,9 @@ Fixed bugs
 * [commit ab8d412c6]: Fix compilation error in CP-violating models
   where the pole mass of a scalar particle is calculated at
   ``LowPrecision``.
+
+* [commit 8a93671da]: Fix non-convergence of QedQcd::to(scale)
+  function when scale is greater than the top pole mass.
 
 
 FlexibleSUSY 2.5.0 [June, 20 2020]